make LoadBalancerWithFacilitiesFactory back to interface
