+++ d2/src/main/java/com/linkedin/d2/xds/balancer/XdsLoadBalancerWithFacilitiesFactory.java
@@ -16,6 +16,7 @@
 
 package com.linkedin.d2.xds.balancer;
 
+import com.linkedin.d2.balancer.AbstractLoadBalancerWithFacilitiesFactory;
 import com.linkedin.d2.balancer.D2ClientConfig;
 import com.linkedin.d2.balancer.LoadBalancerWithFacilities;
 import com.linkedin.d2.balancer.LoadBalancerWithFacilitiesFactory;
@@ -36,7 +37,7 @@ import org.apache.commons.lang3.ObjectUtils;
  * Implementation of {@link LoadBalancerWithFacilitiesFactory} interface, which creates
  * an instance of {@link XdsLoadBalancer}.
  */
-public class XdsLoadBalancerWithFacilitiesFactory extends LoadBalancerWithFacilitiesFactory
+public class XdsLoadBalancerWithFacilitiesFactory extends AbstractLoadBalancerWithFacilitiesFactory
 {
   @Override
   public LoadBalancerWithFacilities create(D2ClientConfig config)
