--- gradle.properties.orig
@@ -1,10 +1,0 @@
-<<<<<<< HEAD
-version=29.67.1
-=======
-version=29.68.0
->>>>>>> deee016b1 (detect LI raw d2 client builder usage)
-group=com.linkedin.pegasus
-org.gradle.configureondemand=true
-org.gradle.parallel=true
-org.gradle.jvmargs=-Xmx4096M
-aggregateFailures=false
